move to component/utils
